requirements.txt
# requirements.txt
python-telegram-bot==20.7
nltk==3.8.1
python-dotenv==1.0.0
pandas==2.1.4
scikit-learn==1.3.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
flask==3.0.0
gunicorn==22.0.0